fix: keep molecular charges float32 on the way to the model
`pad_from_mols` builds the charge tensor float32 and says why at the
construction site. One call frame later `ensemble_opt` re-cast it to
`torch.long`, and the AIMNet2 adapter cast it back to float on arrival.

The round trip survived only because a formal charge is an integer. Any
non-integral charge -- a fragment charge, an ensemble-averaged charge, a
future engine taking a real-valued one -- was truncated toward zero in
silence, and the documented "charges reach a model as float32" contract was
false for the `main()` path.

+++ src/Auto3D/batch_opt/batchopt.py
@@ -100,10 +100,14 @@ def ensemble_opt(
         numbers = torch.tensor(numbers, dtype=torch.long, device=device)
     else:
         numbers = numbers.detach().to(dtype=torch.long, device=device)
+    # float32, not long: `pad_from_mols` builds this tensor float32 on purpose
+    # (see the comment at its construction site) and every adapter casts to its
+    # own working dtype on arrival. Narrowing to int64 here contradicted that
+    # contract and silently truncated any non-integral charge toward zero.
     if not isinstance(charges, torch.Tensor):
-        charges = torch.tensor(charges, dtype=torch.long, device=device)
+        charges = torch.tensor(charges, dtype=torch.float32, device=device)
     else:
-        charges = charges.detach().to(dtype=torch.long, device=device)
+        charges = charges.detach().to(dtype=torch.float32, device=device)
     if atom_mask is not None:
         if not isinstance(atom_mask, torch.Tensor):
             atom_mask = torch.tensor(atom_mask, dtype=torch.bool, device=device)
